Move layouts to their own module.
