+++ backend.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-import abc
 import io
 import os
 import pathlib
 import posixpath
-import re
 import tarfile
 import textwrap
 import time
@@ -16,132 +14,8 @@ from jaraco.functools import pass_none
 from wheel.wheelfile import WheelFile
 
 from . import flit
+from . import layouts
 from .metadata import Message
-
-
-class Layout(abc.ABC):
-    """
-    Lay out tar- and zip-info objects for inclusion in different distributions.
-    """
-
-    def __init__(self, metadata: Message):
-        """
-        Initialize the filter with the root of the package ("coherent/build").
-        """
-        self.metadata = metadata
-
-    @abc.abstractmethod
-    def prefix(self, name: str) -> str:
-        """
-        Given the name, give its location.
-        """
-
-    def __call__(self, info):
-        """
-        Determine disposition for the info object.
-
-        Given an object like a tarfile.TarInfo object, determine if it
-        should be included or filtered. Return None if the object should
-        be omitted. Otherwise, relocate the object to self.prefix.
-        """
-        info.name = info.name.removeprefix('./')
-        ignore_pattern = '|'.join(self.ignored)
-        if info.name != '.' and re.match(ignore_pattern, info.name):
-            return
-        info.name = str(pathlib.PurePosixPath(self.prefix(info.name), info.name))
-        return info
-
-
-class SDist(Layout):
-    """
-    >>> sf = SDist(metadata=types.SimpleNamespace(id="foo"))
-
-    Ignores the .git directory
-    >>> sf(types.SimpleNamespace(name='./.git'))
-
-    Ignores __pycache__ directories
-    >>> sf(types.SimpleNamespace(name='./bar/__pycache__'))
-
-    Ignore paths starting with a dot
-    >>> sf(types.SimpleNamespace(name='./bar/.DS_Store'))
-
-    Ignore dist dirs
-    >>> sf(types.SimpleNamespace(name='./dist'))
-
-    Should not ignore nested dist dirs
-    >>> sf(types.SimpleNamespace(name='./bar/dist'))
-    namespace(name='foo/bar/dist')
-
-    Should not ignore paths that begin with 'dist'
-    >>> sf(types.SimpleNamespace(name='./distributions'))
-    namespace(name='foo/distributions')
-    """
-
-    ignored = ['dist$', r'(.*[/])?__pycache__$', r'(.*[/])?[.]']
-
-    def prefix(self, name):
-        return self.metadata.id
-
-
-class FlitSDist(SDist):
-    """
-    Customize the handling to generate a flit-compatible layout.
-
-    Puts README in the root, but the rest in the package.
-
-    >>> md = Message((('Name', 'foo'), ('Version', '1.0')))
-
-    >>> sf = FlitSDist(metadata=md)
-
-    >>> sf(types.SimpleNamespace(name='./bar.py'))
-    namespace(name='foo-1.0/foo/bar.py')
-
-    >>> sf(types.SimpleNamespace(name='./README.md'))
-    namespace(name='foo-1.0/README.md')
-    """
-
-    ignored = SDist.ignored + [re.escape('pyproject.toml')]
-
-    def prefix(self, name):
-        package = self.metadata['Name'].replace('.', '/')
-        root_pattern = '|'.join(Wheel.ignored)
-        if re.match(root_pattern, name):
-            return pathlib.PurePath(self.metadata.id)
-        return pathlib.PurePath(self.metadata.id, package)
-
-
-class Wheel(Layout):
-    """
-    >>> wf = Wheel(metadata=dict(Name="foo"))
-
-    Ignore all the things SDist does (coherent-oss/coherent.build#33)
-    >>> wf(types.SimpleNamespace(name='./.git'))
-    >>> wf(types.SimpleNamespace(name='./bar/__pycache__'))
-    >>> wf(types.SimpleNamespace(name='./bar/.DS_Store'))
-    >>> wf(types.SimpleNamespace(name='./dist'))
-    >>> wf(types.SimpleNamespace(name='./bar/dist'))
-    namespace(name='foo/bar/dist')
-    >>> wf(types.SimpleNamespace(name='./distributions'))
-    namespace(name='foo/distributions')
-
-    Additionally, filters out non-project files:
-    >>> wf(types.SimpleNamespace(name='./README.rst'))
-    >>> wf(types.SimpleNamespace(name='./docs'))
-    >>> wf(types.SimpleNamespace(name='./(meta)'))
-    >>> wf(types.SimpleNamespace(name='./pyproject.toml'))
-    """
-
-    ignored = SDist.ignored + [
-        'docs',
-        'tests',
-        r'README.*',
-        'PKG-INFO',
-        re.escape('(meta)'),
-        re.escape('pyproject.toml'),
-    ]
-
-    def prefix(self, name):
-        return self.metadata['Name'].replace('.', '/')
 
 
 class ZipInfo(types.SimpleNamespace):
@@ -154,7 +28,7 @@ class ZipInfo(types.SimpleNamespace):
         super().__init__(path=path, name=zip_name)
 
 
-def wheel_walk(filter_: Wheel) -> Iterator[ZipInfo]:
+def wheel_walk(filter_: layouts.Wheel) -> Iterator[ZipInfo]:
     """
     Walk the current directory, applying and honoring the filter for traversal.
     """
@@ -202,7 +76,7 @@ def build_wheel(wheel_directory, config_settings=None, metadata_directory=None):
     )
     filename = pathlib.Path(wheel_directory) / f'{metadata.id}-py3-none-any.whl'
     with WheelFile(filename, 'w') as zf:
-        for info in wheel_walk(Wheel(metadata)):
+        for info in wheel_walk(layouts.Wheel(metadata)):
             zf.write(info.path, arcname=info.name)
         for name, contents in metadata.render_wheel():
             zf.writestr(f'{metadata.id}.dist-info/{name}', contents)
@@ -213,7 +87,7 @@ def build_sdist(sdist_directory, config_settings=None):
     metadata = Message.discover()
     filename = pathlib.Path(sdist_directory) / f'{metadata.id}.tar.gz'
     with tarfile.open(filename, 'w:gz') as tf:
-        tf.add(pathlib.Path(), filter=FlitSDist(metadata))
+        tf.add(pathlib.Path(), filter=layouts.FlitSDist(metadata))
         tf.addfile(*make_sdist_metadata(metadata))
         tf.addfile(*make_flit_project(metadata))
     return filename.name
